requirements.txt
blinker==1.3
Flask==0.10.1
Flask-Admin==1.1.0
Flask-DebugToolbar==0.10.0
Flask-SocketIO==0.6.0
Flask-SQLAlchemy==2.0
gevent==1.0.2
gevent-socketio==0.3.6
gevent-websocket==0.9.5
greenlet==0.4.7
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
RPi.GPIO==0.5.11
SQLAlchemy==1.0.4
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2

